main/requirements.txt
streamlit
nltk
scikit-learn
PyPDF2
pdfminer.six
python-docx
textract
fitz
pandas